fixed issue with component export issue
