+++ src/components/Archive.jsx
@@ -1,0 +1,10 @@
+function Archive() {
+  return (
+    //   TODO implement function that counts all tracked time
+    <>
+      <h2>These are the Tasks you have finished</h2>
+    </>
+  )
+}
+
+export { Archive }
