+++ src/components/Random.jsx
@@ -1,0 +1,9 @@
+function Random() {
+  return (
+    <>
+      <h2>Do this - NOW!!!</h2>
+    </>
+  )
+}
+
+export { Random }
