+++ src/components/styledComponents.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const Body = styled.body`
+  height: 90vh;
+  background: #282c34;
+  padding: 10vw;
+`
+// TODO how to export components
+export { Body }
